chore: redo project init
