--- .eslintrc.yml
@@ -1,12 +1,0 @@
-env:
-  commonjs: true
-  es6: true
-  node: true
-extends:
-  - airbnb-base
-globals:
-  Atomics: readonly
-  SharedArrayBuffer: readonly
-parserOptions:
-  ecmaVersion: 2018
-rules: {}
