+++ .gitignore
@@ -60,8 +60,6 @@ typings/
 # next.js build output
 .next
 
-# WebStorm config folder
 .idea
 
-# build folder
 dist
