+++ package.json
@@ -3,11 +3,10 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "start:dev": "DEBUG=wayfarer-api:*  nodemon --exec npm run babel-node -- ./app",
-    "start": "npm run build && DEBUG=wayfarer-api:* node ./dist/app",
-    "test": "mocha",
-    "babel-node": "babel-node --presets=@babel/env",
-    "build": "babel ./src -d dist --ignore node_modules"
+    "build": "babel ./src -d dist --ignore node_modules",
+    "dev": "DEBUG=wayfarer-api:* nodemon --exec babel-node --presets=@babel/env -- src",
+    "start": "npm run build && DEBUG=wayfarer-api:* node dist",
+    "test": "mocha"
   },
   "dependencies": {
     "cookie-parser": "~1.4.4",
@@ -17,13 +16,13 @@
     "morgan": "~1.9.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.4.5",
-    "@babel/preset-env": "^7.4.5",
     "@babel/cli": "^7.4.4",
+    "@babel/core": "^7.4.5",
     "@babel/node": "^7.4.5",
+    "@babel/preset-env": "^7.4.5",
     "chai": "^4.2.0",
     "eslint": "^5.16.0",
-    "eslint-config-airbnb-base": "^13.1.0",
+    "eslint-config-airbnb-base": "^13.2.0",
     "eslint-plugin-import": "^2.18.0",
     "mocha": "^6.1.4",
     "nodemon": "^1.19.1",
