+++ src/app.js
@@ -1,22 +1,16 @@
 import cookieParser from 'cookie-parser';
-import Debug from 'debug';
+
 import { config as envConfig } from 'dotenv';
 import express from 'express';
-import http from 'http';
+
 import logger from 'morgan';
 import path from 'path';
 
 import indexRouter from './routes';
 
-const debug = Debug('wayfarer-api:server');
-
 envConfig();
 
 const app = express();
-
-/**
- * Configure middleWares
- */
 
 app.use(logger('dev'));
 app.use(express.json());
@@ -24,96 +18,6 @@ app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
 
+app.use('/api', indexRouter);
 
-/**
- * Configure routes
- */
-
-app.use('/', indexRouter);
-
-
-/**
- * Normalize a port into a number, string, or false.
- */
-
-function normalizePort(val) {
-  const port = parseInt(val, 10);
-
-  if (Number.isNaN(port)) {
-    // named pipe
-    return val;
-  }
-
-  if (port >= 0) {
-    // port number
-    return port;
-  }
-
-  return false;
-}
-
-
-/**
- * Get port from environment and store in Express.
- */
-
-const port = normalizePort(process.env.PORT || '3000');
-app.set('port', port);
-
-
-/**
- * Create HTTP server.
- */
-
-const server = http.createServer(app);
-
-
-/**
- * Event listener for HTTP server "error" event.
- */
-
-const onError = (error) => {
-  if (error.syscall !== 'listen') {
-    throw error;
-  }
-
-  const bind = typeof port === 'string'
-    ? `Pipe ${port}`
-    : `Port ${port}`;
-
-  // handle specific listen errors with friendly messages
-  switch (error.code) {
-    case 'EACCES':
-      console.error(`${bind} requires elevated privileges`);
-      process.exit(1);
-      break;
-    case 'EADDRINUSE':
-      console.error(`${bind} is already in use`);
-      process.exit(1);
-      break;
-    default:
-      throw error;
-  }
-};
-
-
-/**
- * Event listener for HTTP server "listening" event.
- */
-
-const onListening = () => {
-  const addr = server.address();
-  const bind = typeof addr === 'string'
-    ? `pipe ${addr}`
-    : `port ${addr.port}`;
-  debug(`Listening on ${bind}`);
-};
-
-
-/**
- * Listen on provided port, on all network interfaces.
- */
-
-server.listen(port);
-server.on('error', onError);
-server.on('listening', onListening);
+export default app;
