+++ src/routes/index.js
@@ -3,7 +3,7 @@ import { Router } from 'express';
 const router = Router();
 
 router.get('/', (req, res) => {
-  res.send('Hello World');
+  res.send('Hello world');
 });
 
 export default router;
